[SC-71] refactor : package 구조 정리
