+++ src/main/java/inu/codin/codin/domain/info/professor/service/ProfessorService.java
@@ -1,10 +1,11 @@
-package inu.codin.codin.domain.info.professor;
+package inu.codin.codin.domain.info.professor.service;
 
 import inu.codin.codin.common.Department;
+import inu.codin.codin.domain.info.professor.domain.Professor;
 import inu.codin.codin.domain.info.professor.dto.ProfessorListResDTO;
 import inu.codin.codin.domain.info.professor.dto.ProfessorThumbnailResDTO;
+import inu.codin.codin.domain.info.professor.repository.ProfessorRepository;
 import lombok.RequiredArgsConstructor;
-import org.bson.types.ObjectId;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
